+++ czxing/src/main/java/me/devilsen/czxing/code/NativeSdk.java
@@ -35,7 +35,7 @@ class NativeSdk {
      * @param points      定位点的位置
      * @param scanType    扫码类型，用于还原定位点的位置
      */
-    public void onDecodeCallback(String content, int formatIndex, float[] points, int scanType) {
+    public void onDecodeCallback(String content, int formatIndex, int[] points, int scanType) {
         if (readCodeListener != null) {
             readCodeListener.onReadCodeResult(new CodeResult(content, formatIndex, points, scanType));
         }
